requirements.txt
python-docx==1.1.0
PyMuPDF==1.23.8
pdfplumber==0.10.3
Flask==3.0.0
Pillow==10.1.0
reportlab==4.0.8
